// src/app/openapi/model/orderItem.ts
/**
 * SpeedyGo API
 * Navigators Team API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface OrderItem { 
    productId?: string;
    quantity?: number;
    price?: number;
}