// src/app/openapi/model/product.ts
/**
 * SpeedyGo API
 * Navigators Team API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Product { 
    id?: string;
    name?: string;
    description?: string;
    price?: number;
    stockQuantity?: number;
    category?: Product.CategoryEnum;
    image?: string;
}
export namespace Product {
    export type CategoryEnum = 'FOOD' | 'CLOTHES' | 'ELECTRONICS' | 'FRAGILEITEM' | 'OTHER';
    export const CategoryEnum = {
        Food: 'FOOD' as CategoryEnum,
        Clothes: 'CLOTHES' as CategoryEnum,
        Electronics: 'ELECTRONICS' as CategoryEnum,
        Fragileitem: 'FRAGILEITEM' as CategoryEnum,
        Other: 'OTHER' as CategoryEnum
    };
}